# Add telemetry trace