--- a/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbConnectionPoolManager.cs
+++ b/src/FirebirdSql.Data.FirebirdClient/FirebirdClient/FbConnectionPoolManager.cs
@@ -167,6 +167,10 @@ sealed class FbConnectionPoolManager : IDisposable
 			var ticks = Environment.TickCount;
 			return ticks + -(long)int.MinValue;
 		}
+
+		internal int AvailableCount => _available.Count;
+		internal int BusyCount => _busy.Count;
+		internal int MaxSize => _connectionString.MaxPoolSize;
 	}
 
 	int _disposed;
@@ -219,6 +223,12 @@ sealed class FbConnectionPoolManager : IDisposable
 			pool.ClearPool();
 		}
 	}
+
+	internal Dictionary<string, (int idleCount, int busyCount, int maxSize)> GetMetrics() =>
+		_pools.ToDictionary(
+			kvp => kvp.Key,
+			kvp => (kvp.Value.AvailableCount, kvp.Value.BusyCount, kvp.Value.MaxSize)
+		);
 
 	public void Dispose()
 	{