… docs/otel.md
… src/FirebirdSql.Data.FirebirdClient/Trace/FbTelemetry.cs