Initial work for #1196.