@fdcastel 2024-10-26
Included the final portion of #1196 (metrics) here, as it requires the same infrastructure as traces.