@willibrandon 2024-10-26
Here is before and after comparison of Perf benchmark results on my machine.
``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 11 (10.0.26120.2130)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=9.0.100-rc.2.24474.11
  [Host]  : .NET 8.0.10 (8.0.1024.46610), X64 RyuJIT AVX2
  NuGet   : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  Project : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2

Jit=RyuJit  Platform=X64  Toolchain=.NET 6.0  
WarmupCount=3  
```
## Perf.CommandBenchmark
FirebirdSQL:master (Old) <-> fdcastel:add-telemetry-trace (**New**)
| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Execute|40,075.4 μs|1,450.87 μs|313.71 KB|
| **New** |	| **34,322.0 μs (-14%)** | **658.70 μs** | **313.11 KB (0%)** |
| Old |Execute|34,465.9 μs|680.65 μs|311.3 KB|
| **New** |	| **34,248.1 μs (-1%)** | **384.64 μs** | **315.91 KB (+1%)** |
| Old |Fetch|640.1 μs|11.02 μs|56.88 KB|
| **New** |	| **648.4 μs (+1%)** | **12.19 μs** | **56.88 KB (0%)** |
| Old |Fetch|653.2 μs|10.90 μs|51.43 KB|
| **New** |	| **653.1 μs (0%)** | **9.41 μs** | **56.06 KB (+9%)** |
| Old |Execute|33,690.3 μs|648.64 μs|319.86 KB|
| **New** |	| **33,559.9 μs (0%)** | **664.54 μs** | **319.85 KB (0%)** |
| Old |Execute|32,943.5 μs|235.67 μs|317.28 KB|
| **New** |	| **34,278.5 μs (+4%)** | **683.34 μs** | **322.01 KB (+1%)** |
| Old |Fetch|642.2 μs|3.47 μs|60.75 KB|
| **New** |	| **647.3 μs (+1%)** | **12.89 μs** | **60.75 KB (0%)** |
| Old |Fetch|664.2 μs|7.49 μs|55.31 KB|
| **New** |	| **658.7 μs (-1%)** | **8.82 μs** | **59.92 KB (+8%)** |