@fdcastel 2024-10-26
Thanks @willibrandon. This surely can be improved.

Could you kindly share the steps/configuration you used for the above benchmark results?

Mine here produces a comparison between "Release" (built from local source) and "ReleaseNuget" (nuget package for FirebirdCliend v8.5.4).