@fdcastel 2025-12-10
@cincuranet Do you wish to move this work forward?

If yes, I could rebase it with the latest `master`.  Along with #1203, which is based on this one