@fdcastel 2026-03-27
For anyone interested: I’ve generated some unofficial packages of the .NET Provider (including this PR) available at:

* [https://github.com/fdcastel/NETProvider/pkgs/nuget/FirebirdSql.Data.FirebirdClient](https://github.com/fdcastel/NETProvider/pkgs/nuget/FirebirdSql.Data.FirebirdClient)
* [https://github.com/fdcastel/NETProvider/pkgs/nuget/FirebirdSql.EntityFrameworkCore.Firebird](https://github.com/fdcastel/NETProvider/pkgs/nuget/FirebirdSql.EntityFrameworkCore.Firebird)

@willibrandon